Change InsLoader to Abstract class, Implement DefaultLoader

+++ src/com/inspedio/entity/basic/InsImage.java
@@ -110,13 +110,13 @@ public class InsImage{
 	public void drawFrameRegion(Graphics g, int Frame, int PercentageX, int PercentageY, int X, int Y, InsAlignment alignment, TransformType transform, InsAlignment Stretch){
 		if(Frame >= 0){
 			Frame = Frame % this.frameTotal;
-			int x_src = ((Frame % this.frameCountX) * this.frameWidth) + ((Stretch.horizontal.getValue() * PercentageX * this.frameWidth) / 200);
-			int y_src = (((int) (Frame / this.frameCountX)) * this.frameHeight) + ((Stretch.vertical.getValue() * PercentageY * this.frameHeight) / 200);
+			int x_src = ((Frame % this.frameCountX) * this.frameWidth) + (int) ((Stretch.horizontal.getValue() * PercentageX * this.frameWidth) / 200);
+			int y_src = (((int) (Frame / this.frameCountX)) * this.frameHeight) + (int) ((Stretch.vertical.getValue() * PercentageY * this.frameHeight) / 200);
 			this.drawRegion(g, x_src, y_src, (this.frameWidth * PercentageX) / 100, (this.frameHeight * PercentageY) / 100, transform, X, Y, alignment);
 		} else {
 			Frame = InsUtil.Absolute(Frame) % this.frameTotal;
-			int x_src = ((Frame % this.frameCountX) * this.frameWidth) + ((Stretch.horizontal.getValue() * PercentageX * this.frameWidth) / 200);
-			int y_src = (((int) (Frame / this.frameCountX)) * this.frameHeight) + ((Stretch.vertical.getValue() * PercentageY * this.frameHeight) / 200);
+			int x_src = ((Frame % this.frameCountX) * this.frameWidth) + (int) ((Stretch.horizontal.getValue() * PercentageX * this.frameWidth) / 200);
+			int y_src = (((int) (Frame / this.frameCountX)) * this.frameHeight) + (int) ((Stretch.vertical.getValue() * PercentageY * this.frameHeight) / 200);
 			this.drawRegion(g, x_src, y_src, (this.frameWidth * PercentageX) / 100, (this.frameHeight * PercentageY) / 100, TransformType.getMirror(transform), X, Y, alignment);
 		}
 	}
